# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ollama Node API",
    "version": "1.1.0",
    "description": "API experimental para tarefas curtas. Clique em Authorize e cole somente API_KEY (sem o prefixo Bearer). No Render Free, a primeira chamada pode demorar pelo cold start. Até 8 KB por JSON."
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesmo servidor desta documentação"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "Status"
    },
    {
      "name": "Inferência"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cole apenas o valor de API_KEY."
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Status"
        ],
        "operationId": "health",
        "summary": "Verificar processo Node",
        "security": [],
        "responses": {
          "200": {
            "description": "Processo ativo; não testa inferência.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "alive"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": [
          "Status"
        ],
        "operationId": "ready",
        "summary": "Verificar Ollama e modelo em disco",
        "description": "Não comprova RAM suficiente para inferência.",
        "responses": {
          "200": {
            "description": "Modelo presente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "model_present"
                      ]
                    },
                    "note": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "note"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Bearer ausente ou incorreto.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Ollama indisponível ou modelo ausente.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "error"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "not_ready"
                          ]
                        },
                        "note": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "status",
                        "note"
                      ]
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/classify": {
      "post": {
        "tags": [
          "Inferência"
        ],
        "operationId": "classify",
        "summary": "Classificar texto entre 2 ou 3 categorias",
        "description": "Uma inferência por vez. Contexto de 512 tokens. Até 16 tokens na saída; a categoria validada pode estar semanticamente errada.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 600,
                    "description": "Texto curto, não pode conter apenas espaços. Limite do servidor medido em unidades UTF-16."
                  },
                  "categories": {
                    "type": "array",
                    "minItems": 2,
                    "maxItems": 3,
                    "uniqueItems": true,
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 24
                    },
                    "description": "Categorias distintas também sem diferenciar maiúsculas. Somente letras Unicode, números, hífen ou sublinhado."
                  }
                },
                "required": [
                  "text",
                  "categories"
                ]
              },
              "example": {
                "text": "Gostei muito do atendimento!",
                "categories": [
                  "elogio",
                  "reclamacao",
                  "duvida"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado gerado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "category": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "category",
                    "model"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Bearer ausente ou incorreto.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que 8 KB.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Use application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Modelo retornou categoria inválida; raw contém a saída.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Modelo ocupado. Aguarde antes de tentar novamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Segundos até nova tentativa.",
                "schema": {
                  "type": "integer",
                  "example": 5
                }
              }
            }
          },
          "502": {
            "description": "Resposta inválida do Ollama.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Ollama indisponível ou falha de inferência.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "504": {
            "description": "Tempo limite de inferência excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/generate": {
      "post": {
        "tags": [
          "Inferência"
        ],
        "operationId": "generate",
        "summary": "Gerar resposta curta",
        "description": "Uma inferência por vez. Contexto de 512 tokens. Até 30 tokens na saída; resposta pode ser cortada. Sem histórico.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 600,
                    "description": "Texto curto, não pode conter apenas espaços. Limite do servidor medido em unidades UTF-16."
                  }
                },
                "required": [
                  "prompt"
                ]
              },
              "example": {
                "prompt": "Crie um título curto para uma promoção de frutas."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado gerado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "response",
                    "model"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Bearer ausente ou incorreto.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que 8 KB.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Use application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Modelo ocupado. Aguarde antes de tentar novamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Segundos até nova tentativa.",
                "schema": {
                  "type": "integer",
                  "example": 5
                }
              }
            }
          },
          "502": {
            "description": "Resposta inválida do Ollama.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Ollama indisponível ou falha de inferência.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "504": {
            "description": "Tempo limite de inferência excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Erro interno.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "raw": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}